feat: complete all 7 MVs + 2 risk views, stepwise approach for heavy queries
- mv_servidor_pb_risco: fix BOOL_OR(EXISTS) anti-pattern with stepwise temp tables
  (correlated subquery per group: 26s→10min+ → stepwise: ~3min total)
- mv_rede_pb: stepwise edge-type tables to avoid 30min+ UNION ALL timeout
- mv_empresa_pb: 157k rows PB-focused company profiles
- Composite indexes: idx_socio_norm_nome (1040MB), idx_bf_cpf_nome (910MB)
- ETL 04b_pncp_itens.py rewritten: thread pool + COPY (5x faster)
- New: etl/21_views.py (statement-by-statement executor), CI/CD skeleton
- Fix: pncp_item numero_item SMALLINT→INTEGER overflow

+++ sql/03b_schema_pncp_itens.sql
@@ -6,7 +6,7 @@ DROP TABLE IF EXISTS pncp_item CASCADE;
 
 CREATE TABLE pncp_item (
     numero_controle_pncp     VARCHAR(50),        -- FK para pncp_contratacao
-    numero_item              SMALLINT,
+    numero_item              INTEGER,
     descricao                TEXT,
     material_ou_servico      CHAR(1),             -- M=Material, S=Servico
     valor_unitario_estimado  DECIMAL(15,2),
